LayoutTests/editing/execCommand/format-block-typing-style.html

<!DOCTYPE html>
<html>
<body>
<div id="test" contenteditable>hello</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description("This tests ensures FormatBlock preserves the typing style.");

var test = document.getElementById('test');
window.getSelection().setPosition(test, 1);

document.execCommand('Bold', false, null);
document.execCommand('FormatBlock', false, '');
document.execCommand('InsertText', false, ' world');

Markup.dump(test, '"world" should be bolded');

</script>
</body>
</html>
