third_party/blink/web_tests/platform/generic/editing/execCommand/format-block-typing-style-expected.txt

This tests ensures FormatBlock preserves the typing style.

"world" should be bolded:
| "hello"
| <b>
|   " world<#selection-caret>"
